Add configuration file (configuration.json)
Initial configuration is done through the JSON file configuration.json.
It's used to control if there is an admin user in the system.

Maybe there's a better way. For now it should work.

It's useful to reduce the total of lines that need to be modified in the
app.rb file.